refactor(frontend): sqlIdentifier の statement type detection を utils/sql/statement-type.ts (操作層) に分離
